clean up argument changes, fix DRY fails, add stat tests
