+++ dockers/Cas9Emulation.py
@@ -1,3 +1,47 @@
+from classes.Cas9 import Cas9
+
+def convert_cas9_to_tuple(key: int, guide: Cas9) -> (int, str, str, str, str, any, any):
+    return (key,
+            guide.downstream_5_prim,
+            guide.downstream_3_prim,
+            guide.stranded_guide_seq,
+            guide.pam,
+            guide.repair_profile,
+            guide.repair_stats)
+
+def recv_tuples():
+    """
+    Receives a Base64 encoded pickled list of tuples containing arguments for Cas9Emulation objects.
+
+    Source: https://stackoverflow.com/q/30469575
+
+    :return a list of tuples
+    """
+    pickled = sys.stdin.read()
+
+    tuples = pickle.loads(codecs.decode(pickled.encode(), 'base64'))
+
+    return tuples
+
+def tuple_to_cas9(t):
+    """
+    Takes a tuple containing all fields of the Cas9Emulation object and returns a Cas9Emulation object.
+
+    :param t: The tuple representation of a Cas9Emulation object.
+    :return: The corresponding Cas9Emulation object.
+    """
+    return Cas9Emulation.Cas9Emulation(*t)
+
+def cas9_to_reduced_tuple(guide):
+    """
+    Returns a reduced tuple containing only the key & modified values of the Cas9Emulation objects after running
+    repair prediction.
+
+    :param guide: The guide to reduce into a tuple.
+    :return: A tuple containing the inout guide's key, score
+    """
+    return guide.key, guide.repProfile, guide.repStats
+
 class Cas9Emulation(object):
     """
     `Cas9` emulation.
@@ -19,4 +63,3 @@ class Cas9Emulation(object):
         self.PAM = pam.encode('ascii', 'ignore')
         self.repProfile = rep_profile
         self.repStats = rep_stats
-
