+++ dockers/li_2021_wrapper.py
@@ -4,29 +4,20 @@ import subprocess
 
 import config
 from classes.Cas9 import Cas9
-
-
-def convert_cas9_to_tuple(key: int, guide: Cas9) -> (int, str, str, str, str, any, any):
-    return (key,
-            guide.downstream_5_prim,
-            guide.downstream_3_prim,
-            guide.stranded_guide_seq,
-            guide.pam,
-            guide.repair_profile,
-            guide.repair_stats)
-
+from dockers import Cas9Emulation as c9
 
 def run_croton_prediction(repair_prediction: str, guides: [Cas9]) -> [Cas9]:
     """
-    Runs chopchop_inDelphi docker image using the supplied guides & repair prediction mode.
+    Runs chopchop_li_2021 docker image using the supplied guides & repair prediction mode.
 
-    :param repair_prediction: The repair mode to use. Accepted values are 'mESC', 'U2OS', 'HEK293', 'HCT116' & 'K562'.
+    :param repair_prediction: The repair mode to use.
     :param guides: A list of Cas9 objects to score.
     :return: Returns a list of Cas9 scored objects.
     """
+
     keyed_tuples = []
     for key, guide in enumerate(guides):
-        keyed_tuples.append(convert_cas9_to_tuple(key, guide))
+        keyed_tuples.append(c9.convert_cas9_to_tuple(key, guide))
 
     encoded = codecs.encode(pickle.dumps(keyed_tuples, protocol=2), 'base64').decode()
