+++ testing/end_to_end/test_data/HCT116_repStats.json
@@ -1,0 +1,1 @@
+{"Expected indel length": 5.446130139484777, "Phi": 2.678901093221185, "Frame +0 frequency": 45.6180761841144, "Frameshift frequency": 54.38192381588569, "Reference sequence": "ATTTACGGCAAGCAACACAGCTCGTAGGTTTATAAAGGGATGGGGTGTAGGAATTCACGGGCGCGGTCAGCCGAGCTCAGAAGCATTGGATTGCCTAATTGCGCGCGGTAAATGCCTTGC", "gRNA": null, "Highest del frequency": 32.950383089552886, "Precision": 0.5458683048480671, "Cutsite": 60, "Highest ins frequency": 6.504530256027554, "Frame +1 frequency": 25.718061491481535, "MH del frequency": 77.62335496230914, "1-bp ins frequency": 13.108927289049966, "gRNA orientation": null, "Cas9 type": null, "MHless del frequency": 9.267717748641031, "Frame +2 frequency": 28.663862324404157, "Highest outcome frequency": 32.950383089552886}
