+++ testing/end_to_end/test_data/HEK293_repStats.json
@@ -1,0 +1,1 @@
+{"Expected indel length": 5.6973995840860425, "Phi": 2.678901093221185, "Frame +0 frequency": 48.19614482064977, "Frameshift frequency": 51.80385517935038, "Reference sequence": "ATTTACGGCAAGCAACACAGCTCGTAGGTTTATAAAGGGATGGGGTGTAGGAATTCACGGGCGCGGTCAGCCGAGCTCAGAAGCATTGGATTGCCTAATTGCGCGCGGTAAATGCCTTGC", "gRNA": null, "Highest del frequency": 34.81254730845046, "Precision": 0.5524067648905184, "Cutsite": 60, "Highest ins frequency": 3.745776240161357, "Frame +1 frequency": 21.520077886102197, "MH del frequency": 82.0101760128792, "1-bp ins frequency": 8.198348674403794, "gRNA orientation": null, "Cas9 type": null, "MHless del frequency": 9.791475312717214, "Frame +2 frequency": 30.283777293248185, "Highest outcome frequency": 34.81254730845046}
