+++ testing/end_to_end/test_data/K562_repStats.json
@@ -1,0 +1,1 @@
+{"Expected indel length": 5.618201630620063, "Phi": 2.678901093221185, "Frame +0 frequency": 47.38355990714222, "Frameshift frequency": 52.616440092857914, "Reference sequence": "ATTTACGGCAAGCAACACAGCTCGTAGGTTTATAAAGGGATGGGGTGTAGGAATTCACGGGCGCGGTCAGCCGAGCTCAGAAGCATTGGATTGCCTAATTGCGCGCGGTAAATGCCTTGC", "gRNA": null, "Highest del frequency": 34.22560926913462, "Precision": 0.5515450878770317, "Cutsite": 60, "Highest ins frequency": 5.487985615743967, "Frame +1 frequency": 22.84324597268487, "MH del frequency": 80.62748799851292, "1-bp ins frequency": 9.746120538312255, "gRNA orientation": null, "Cas9 type": null, "MHless del frequency": 9.626391463174823, "Frame +2 frequency": 29.77319412017304, "Highest outcome frequency": 34.22560926913462}
