+++ testing/end_to_end/test_data/U2OS_repStats.json
@@ -1,0 +1,1 @@
+{"Expected indel length": 5.104026127561426, "Phi": 2.678901093221185, "Frame +0 frequency": 42.10802893196194, "Frameshift frequency": 57.89197106803816, "Reference sequence": "ATTTACGGCAAGCAACACAGCTCGTAGGTTTATAAAGGGATGGGGTGTAGGAATTCACGGGCGCGGTCAGCCGAGCTCAGAAGCATTGGATTGCCTAATTGCGCGCGGTAAATGCCTTGC", "gRNA": null, "Highest del frequency": 30.415041591282275, "Precision": 0.5448715511886754, "Cutsite": 60, "Highest ins frequency": 10.95166546641286, "Frame +1 frequency": 31.433627248662212, "MH del frequency": 71.65068652516021, "1-bp ins frequency": 19.794693031881916, "gRNA orientation": null, "Cas9 type": null, "MHless del frequency": 8.554620442957853, "Frame +2 frequency": 26.458343819375948, "Highest outcome frequency": 30.415041591282275}
