+++ testing/end_to_end/test_data/croton_repStats.json
@@ -1,0 +1,1 @@
+{"Deletion frequency": 90.37785530090332, "1 bp insertion probability": 3.9933115243911743, "1 bp deletion probability": 20.828557014465332, "1 bp frameshift frequency": 27.36549973487854, "2 bp frameshift frequency": 47.91848361492157, "Frameshift frequency": 75.59791207313538}
