+++ testing/end_to_end/test_data/mESC_repStats.json
@@ -1,0 +1,1 @@
+{"Expected indel length": 5.8165564481973036, "Phi": 2.678901093221185, "Frame +0 frequency": 49.41871517607316, "Frameshift frequency": 50.58128482392699, "Reference sequence": "ATTTACGGCAAGCAACACAGCTCGTAGGTTTATAAAGGGATGGGGTGTAGGAATTCACGGGCGCGGTCAGCCGAGCTCAGAAGCATTGGATTGCCTAATTGCGCGCGGTAAATGCCTTGC", "gRNA": null, "Highest del frequency": 35.695621846766045, "Precision": 0.559262969012944, "Cutsite": 60, "Highest ins frequency": 3.514273481613256, "Frame +1 frequency": 19.52931230455138, "MH del frequency": 84.09049198855543, "1-bp ins frequency": 5.869656661645507, "gRNA orientation": null, "Cas9 type": null, "MHless del frequency": 10.039851349799326, "Frame +2 frequency": 31.05197251937561, "Highest outcome frequency": 35.695621846766045}
